Medical_Q&A_Bot/backend/requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
azure-openai